refactor: extract the lazy properties into a mixin
The `# Properties` banner block of PdsFile moves to src/pdsfile/_properties.py as
_PropertiesMixin, the last and largest of the Phase-5 mixins (settled decision
8.3). The moved text is the exact concatenation of two line ranges of the parent's
pdsfile.py, 672-2034 and 2037-2230 -- 1,557 lines, byte-identical, MD5
a49cd66334d4952bd82c6b9b518ce246 -- and the remainder of the class body after the
class statement is byte-identical to the parent's with those two ranges deleted.

Sixty-eight of the block's sixty-nine statements move: 64 lazy properties, the
version_info staticmethod, and the all_versions, viewset_lookup and
_repair_width_height methods. The sixty-ninth is LATEST_VERSION_RANKS, a class
attribute and a frozen manifest member, which stays on PdsFile with the two
alternative constructors that are its only readers. The mixin defines no __init__
and no new state: every attribute it writes is one of the 41 slots PdsFile.__init__
already creates, and _recache and _complete stay in core.

Minimal keep-green edits carried by this commit, each required to keep the package
importable and every active gate green:

  * the _PropertiesMixin import and its alphabetical slot in the class bases;
  * datetime, PIL, pdsparser, pdsviewable and formatted_file_size lose their last
    local reader in pdsfile.py. All five are frozen members of pdsfile.pdsfile
    (tests/api/api_manifest.json), so they move to the redundant-alias re-export
    form rather than being deleted;
  * the ruff ratchet entry for _properties.py, a split of pdsfile.py's: E701,
    F841, RUF005, SIM114, UP024, UP031, every code already in pdsfile.py's entry.
    All 17 of pdsfile.py's codes conserve exactly across the split, and RUF005
    leaves pdsfile.py entirely, so pdsfile.py's entry drops it: 17 codes to 16.
